+++ NEWS.md
@@ -3,8 +3,18 @@ title: "Release notes for `PFUDatabase`"
 output: html_document
 ---
 
+Cite all releases with doi [10.5281/zenodo.5228375](https://doi.org/10.5281/zenodo.5228375), 
+which always resolves to the latest release.
 
-# PFUDatabase 0.1.1 (2023-05-18)
+
+# PFUDatabase 0.1.2 (2023-08-18)
+
+* Adapted to changes in `MWTools` for downloading and
+  analyzing ILO data.
+* New targets `PSUT_USA` and `EtafuYEIOU`.
+
+
+# PFUDatabase 0.1.1 (2023-05-18) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7949498.svg)](https://doi.org/10.5281/zenodo.7949498)
 
 * Released version 1.1 of the database.
   So also releasing this version of the package.